Assignments/HW_2/requirements.txt
Flask==1.1.2
flask-restplus==0.13.0
Keras==2.2.4
numpy==1.19.1
Pillow==7.2.0
tensorflow==1.14.0
Werkzeug==1.0.1